chore: fix namings
