+++ node/p2p/routing.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 
 	"github.com/ipfs/go-datastore"
-	"github.com/libp2p/go-libp2p-core/peer"
 	"github.com/libp2p/go-libp2p-core/protocol"
 	"github.com/libp2p/go-libp2p-core/routing"
 	dht "github.com/libp2p/go-libp2p-kad-dht"
@@ -26,13 +25,9 @@ func ContentRouting() routing.ContentRouting {
 // Basically, this provides a way to discover peer addresses by respecting public keys.
 func PeerRouting(cfg Config) func(routingParams) (routing.PeerRouting, error) {
 	return func(params routingParams) (routing.PeerRouting, error) {
-		bpeers := make([]peer.AddrInfo, len(params.Peers))
-		for index := range params.Peers {
-			bpeers[index] = *params.Peers[index]
-		}
 		opts := []dht.Option{
 			dht.Mode(dht.ModeAuto),
-			dht.BootstrapPeers(bpeers...),
+			dht.BootstrapPeers(params.Peers...),
 			dht.ProtocolPrefix(protocol.ID(fmt.Sprintf("/celestia/%s", params.Net))),
 			dht.Datastore(params.DataStore),
 			dht.QueryFilter(dht.PublicQueryFilter),
@@ -72,7 +67,7 @@ type routingParams struct {
 
 	Ctx       context.Context
 	Net       nparams.Network
-	Peers     nparams.BootstrapPeers
+	Peers     nparams.Bootstrappers
 	Lc        fx.Lifecycle
 	Host      HostBase
 	DataStore datastore.Batching
