+++ node/tests/p2p.go
@@ -1,0 +1,39 @@
+package tests
+
+import (
+	"context"
+	"testing"
+	"time"
+
+	"github.com/libp2p/go-libp2p-core/host"
+	"github.com/libp2p/go-libp2p-core/network"
+	"github.com/libp2p/go-libp2p-core/peer"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/celestiaorg/celestia-node/node"
+	"github.com/celestiaorg/celestia-node/node/tests/swamp"
+)
+
+func TestUseBridgeNodeAsBootstraper(t *testing.T) {
+	sw := swamp.NewSwamp(t)
+
+	bridge := sw.NewBridgeNode()
+
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
+	t.Cleanup(cancel)
+
+	err := bridge.Start(ctx)
+	require.NoError(t, err)
+
+	addr := host.InfoFromHost(bridge.Host)
+
+	full := sw.NewFullNode(node.WithBootstrappers([]peer.AddrInfo{*addr}))
+	light := sw.NewLightNode(node.WithBootstrappers([]peer.AddrInfo{*addr}))
+	nodes := []*node.Node{full, light}
+	for index := range nodes {
+		require.NoError(t, nodes[index].Start(ctx))
+		assert.Equal(t, *addr, nodes[index].BootstrapPeers[0])
+		assert.True(t, nodes[index].Host.Network().Connectedness(addr.ID) == network.Connected)
+	}
+}
