+++ params/bootstrap.go
@@ -9,7 +9,7 @@ import (
 var log = logging.Logger("params")
 
 // BootstrappersInfosFor returns address information of bootstrap peers for a given network.
-func BootstrappersInfosFor(net Network) (BootstrapPeers, error) {
+func BootstrappersInfosFor(net Network) (Bootstrappers, error) {
 	bs, err := bootstrappersFor(net)
 	if err != nil {
 		return nil, err
@@ -38,8 +38,8 @@ var bootstrapList = map[Network][]string{
 }
 
 // parseAddrInfos converts strings to AddrInfos
-func parseAddrInfos(addrs []string) ([]*peer.AddrInfo, error) {
-	infos := make([]*peer.AddrInfo, 0, len(addrs))
+func parseAddrInfos(addrs []string) ([]peer.AddrInfo, error) {
+	infos := make([]peer.AddrInfo, 0, len(addrs))
 	for _, addr := range addrs {
 		maddr, err := ma.NewMultiaddr(addr)
 		if err != nil {
@@ -52,7 +52,7 @@ func parseAddrInfos(addrs []string) ([]*peer.AddrInfo, error) {
 			log.Errorw("parsing info from multiaddr", "maddr", maddr, "err", err)
 			return nil, err
 		}
-		infos = append(infos, info)
+		infos = append(infos, *info)
 	}
 
 	return infos, nil
